Refactored to make the Grid rendering configurable
